+++ tests/unit/asset/test_asset_base.py
@@ -6,12 +6,14 @@
 import secrets
 import logging
 import json
+import pytest
 
 from starfish.asset import AssetBase
 from starfish.utils.did import (
     did_to_id,
     id_to_did,
-    decode_to_asset_id
+    decode_to_asset_id,
+    did_generate_random
 )
 
 ASSET_METADATA = {
@@ -19,7 +21,7 @@ ASSET_METADATA = {
     'type': 'asset',
 }
 
-TEST_DID = id_to_did(secrets.token_hex(32))
+TEST_DID = did_generate_random()
 
 def test_init(metadata):
     asset = AssetBase(ASSET_METADATA)
@@ -47,15 +49,13 @@ def test_is_asset_type():
 def test_asset_id():
     asset = AssetBase(ASSET_METADATA, TEST_DID)
     assert(asset)
-    asset_id = decode_to_asset_id(asset.did)
-    print(asset_id, asset.asset_id)
+    with pytest.raises(ValueError):
+        asset_id = decode_to_asset_id(asset.did)
 
-    assert(asset.asset_id == asset_id)
+    test_did = did_generate_random()
+    asset_id = secrets.token_hex(32)
+    asset_did = f'{test_did}/{asset_id}'
 
-    did_id = secrets.token_hex(32)
-    path_did = id_to_did(did_id)
-    asset_id = decode_to_asset_id(f'{path_did}/{asset_id}')
-
-    asset = AssetBase(ASSET_METADATA, f'{path_did}/{asset_id}')
+    asset = AssetBase(ASSET_METADATA, asset_did)
     assert(asset)
     assert(asset.asset_id == asset_id)